requirements.txt
stripe>=7.0.0
google-cloud-bigquery>=3.14.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
